Merge http://lia64.bkbits.net/linux-ia64-release-2.6.11 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2005/02/08 09:59:28-08:00 kaos
[IA64] mca_asm.S: Correctly dereference ia64_mca_data

arch/ia64/kernel/mca_asm.S is treating per_cpu__ia64_mca_data as the
start of the mca data, instead of as a pointer to the mca data.  It
ends up overwriting the rest of the per cpu area with the MCA stack and
bspstore.  Since we dereference ia64_mca_data several times, make it a
macro.

Signed-off-by: Keith Owens <kaos@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

BKrev: 420a7e9aTZDAN-fePc6j7tvwSrptcA