Merge bk://bk.arm.linux.org.uk/linux-2.6-rmk into ppc970.osdl.org:/home/torvalds/v2.6/linux
2005/02/14 14:21:09-08:00 eike-kernel
[PATCH] make ACPI_BLACKLIST_YEAR depend on ACPI_INTERPRETER

this oneliner fixes the situation that I can enter a year to blacklist
ACPI devices even if ACPI is completely disabled.

Signed-off-by: Rolf Eike Beer <eike-kernel@sf-tec.de>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

2005/02/14 14:20:06-08:00 hch
[PATCH] block new writers on frozen filesystems

When the lockfs patches went in an important bit got lost, the call in
generic_file_write to put newly incoming writers to sleep when a filesystem
is frozen.  Nathan added back the call in the now separate XFS write patch,
and the patch for the generic code is below:

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

2005/02/14 14:19:53-08:00 minyard
[PATCH] IPMI: Fix LAN bridging

The size of LAN bridged messages was not being returned properly from the
function that calculated address sizes.

Signed-off-by: Corey Minyard <minyard@acm.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

2005/02/14 14:19:40-08:00 olh
[PATCH] ppc64: remove extra whitespace before preprocessor token

unifdef complains about the space before #ifndef.

Signed-off-by: Olaf Hering <olh@suse.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

2005/02/14 14:19:26-08:00 herbert
[PATCH] ISDN locking fix

isdn_net_get_locked_lp is doing a local_bh_enable with hard IRQs disabled.
This is not allowed.

The following patch fixes the problem by removing the unnecessary
local_bh_enable while the hard IRQs are disabled.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 421145c1mvuz1crh0WAMxEkScUGuUQ