[PATCH] radeonfb: Fix hang on boot with some laptops
It appears that access to the PLL registers of the radeon chip is
unreliable while the card is in "legacy VGA" text mode. I don't have a
good explanation yet, it might be the BIOS mucking around behind my
back. This fixes the lockup by moving the code that enables/disables the
dynamic power management to after the mode is set.

I'm still waiting for a proper explanation from ATI...

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4212b22bxYbMvpz4bU79ob5G4Zps8g