Merge bk://kernel.bkbits.net/gregkh/linux/fix-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2005/02/16 14:27:17-08:00 david-b
[PATCH] USB: ehci requeue revisit

This gets rid of a bug found in some IRQ handling logic, after tripping
a debug assertion.  Basically, a recent patch called the wrong routine to
unlink a QH.  Net result, it wasn't allowing for the case that some other
QH was already being unlinked.  This patch uses the correct routine; the
names are confusingly similar, and the effect is often identical.

The consequence of using the wrong routine was that the driver could
lose one of the pending unlinks (probably wedging some activity) and
treat the other one as completed before it was safe to do so (which
probably wouldn't oops, but could cause other nasty corruption).

From: Brian Murphy <brian@murphy.dk>
Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>

2005/02/16 14:26:53-08:00 stern
[PATCH] USB Hub driver: Add reset recovery-time delay

This patch is clearly needed for us to be in compliance with the USB spec.
It adds the mandated recovery-time delay following a port reset.
Regardless of anything else we do to alter the device initialization
sequence, this is necessary.


Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>

2005/02/16 14:26:30-08:00 david-b
[PATCH] USB: ehci patch for NF4 port miscounting

Turns out that a workaround for a different EHCI chip trips up at
least one NForce4 board.  Neither controller can multiply right.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>

BKrev: 4213d5efo04qCCsFJn-9MQw3T-ryaA