[PATCH] natsemi long cable fix
This is a minor modification to the previous patch submission that does
not assume the default contents of the DSPCFG register are zero.

When used with Revision D of the DP83815, the "Recommended Registers
Configuration" from page 78 of the DP83815 data sheet is not entirely
compatible with the driver's "short cable patch".  When the DSPCFG
register is written with the value suggested in the document, then
do_cable_magic() can't read the DSP coefficient and determines that all
cables attached to the DP83815D are 'short', regardless of actual
length.  Short cables (< 30m) cause do_cable_magic to enable additional
attenuation to reduce CRC and idle errors.  If the extra attenuation is
unintentionally enabled for long cables (> 50m?), they will not operate
properly.  The National Semiconductor driver, 'dp83815.c' from
http://www.national.com/appinfo/networks/files/linux_2_4.tar.gz was used
as a basis for this modification.

Signed-off-by: Jeff Garzik <jgarzik@pobox.com>

BKrev: 42194504MstpFKgwjMiDOZcejgcQTQ