[PATCH] ppc64: fix compilation for Maple board
A patch that I sent in earlier to allow the use of the data address
breakpoint on machines with a hypervisor happened to break things for those
configs, such as for the Maple board, where we don't compile in the
routines for calling the hypervisor.  This patch fixes it.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 421cb240puXjn1KOliqLctt7tdCJPg