Merge bk://bk.arm.linux.org.uk/linux-2.6-serial into ppc970.osdl.org:/home/torvalds/v2.6/linux
2005/02/28 16:09:45+00:00 alex.williamson
[SERIAL] 8250 woraround for buggy uart

Patch from Alex Williamson

This patch adds support for detecting and working around a bug in
the A2 rev of the Exar ST16C2550 UART.  The chip incorrectly
advertises an EFR and mis-detects as having the wrong size FIFO.
Much of the patch below is Russell's proposed solution to the
problem.  The only changes I've made are to check the FIFO size
on the part (because there is a real part with the same divisor
ID and larger FIFO) and save and restore the LCR register around
the size_fifo() routine (it doesn't work correctly with a LCR
value of 0xBF).

Signed-off-by: Alex Williamson
Signed-off-by: Russell King

2005/02/11 23:23:06+00:00 bjorn.helgaas
[SERIAL] add TP560 data/fax/modem support

Patch from Bjorn Helgaas

Claim Topic TP560 data/fax/voice modem.  This device reports as class 0x0780,
so we don't claim it by default:

        00:0d.0 Class 0780: 151f:0000
                Subsystem: 151f:0000
                Interrupt: pin A routed to IRQ 11
                Region 0: I/O ports at a400 [size=8]
        00: 1f 15 00 00 01 00 00 02 00 00 80 07 00 00 00 00
        10: 01 a4 00 00 00 00 00 00 00 00 00 00 00 00 00 00
        20: 00 00 00 00 00 00 00 00 00 00 00 00 1f 15 00 00
        30: 00 00 00 00 00 00 00 00 00 00 00 00 0b 01 00 00

Some rc.serial scripts extract IRQ and I/O port information from
/proc/pci and stuff it into an unused port using setserial.  That
doesn't work reliably anymore because pci_enable_device() is never
called, so the IRQ may not be enabled.

Thanks to Evan Clarke for reporting and helping debug this problem.

Signed-off-by: Bjorn Helgaas

2005/02/11 23:04:52+00:00 c.lucas
[SERIAL] drivers/serial/*: convert to pci_register_driver

Patch from Christophe Lucas

convert from pci_module_init to pci_register_driver

Signed-off-by: Christophe Lucas

2005/02/09 21:18:47+00:00 bjorn.helgaas
[SERIAL] discover PNP ports before PCI, etc

Patch from Bjorn Helgaas

PNP ports tend to be built-in, and discovering them after
PCI ports means the names of the built-in ports can change
if you add or remove PCI ports.

(And yes, we should look at getting rid of 8250_acpi.c
now that we have PNPACPI, but that's for another patch.)

Signed-off-by: Bjorn Helgaas

BKrev: 422507c2xV8Kb_MpneE_K9UrULP1bg