[PATCH] SGI 921857: find broken with nohide on NFSv3
This patch makes "find" work when traversing nohide exports on NFSv3.  The
READDIRPLUS reply needs to not return a file handle for the ".." entry when
the directory is a server side mountpoint, which would be the directory
itself, otherwise the client remembers the wrong file handle and gets
confused.

Signed-off-by: Greg Banks <gnb@melbourne.sgi.com>
Signed-off-by: Neil Brown <neilb@cse.unsw.edu.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4229e896Ae6s9P190XFcQ2Zp62N4mQ