[PATCH] ppc64: collect and export low-level cpu usage statistics
POWER5 machines have a per-hardware-thread register which counts at a rate
which is proportional to the percentage of cycles on which the cpu
dispatches an instruction for this thread (if the thread gets all the
dispatch cycles it counts at the same rate as the timebase register).  This
register is also context-switched by the hypervisor.  Thus it gives a
fine-grained measure of the actual cpu usage by the thread over time.

This patch adds code to read this register every timer interrupt and on
every context switch.  The total over all virtual processors is available
through the existing /proc/ppc64/lparcfg file, giving a way to measure the
total cpu usage over the whole partition.

Signed-off-by: Manish Ahuja <ahuja@austin.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4229ed2f1tv8QCCoN4bPrSS9b1nJ1A